chore: 폴더구조 수정 및 fetch params 수정
