+++ src/App.jsx
@@ -1,24 +1,17 @@
-import { Route, BrowserRouter as Router, Routes, useSearchParams } from "react-router-dom";
+import { Route, BrowserRouter as Router, Routes } from "react-router-dom";
 import { PageEndPoints } from "./constants/api";
 import AuthProvider from "./contexts/AuthProvider";
+import SubscriptionsProvider from "./contexts/SubscriptionsProvider";
 import ChannelPage from "./pages/Channel/Channel";
+import FeedChannelsPage from "./pages/Feed/Channels/Channels";
+import FeedplayListsPage from "./pages/Feed/PlayLists/PlayLists";
+import FeedSubscriptionsPage from "./pages/Feed/Subscriptions/Subscriptions";
 import HomePage from "./pages/Home/Home";
 import LoadingPage from "./pages/Loading/Loading";
-import VideoPlayer from "./pages/VideoPlayer/VideoPlayer";
-import ShortsPlayer from "./pages/ShortsPlayer/ShortsPlayer";
-import FeedplayListsPage from "./pages/Feed/PlayLists/PlayLists";
-import FeedChannelsPage from "./pages/Feed/Channels/Channels";
-import FeedSubscriptionsPage from "./pages/Feed/Subscriptions/Subscriptions";
 import ResultsPage from "./pages/Results/Results";
-import SubscriptionsProvider from "./contexts/SubscriptionsProvider";
-import Like from "./pages/Like/Like";
-import PlayListItem from "./pages/Feed/PlayLists/PlayListItem/PlayListItem";
-
-const PlaylistRoute = () => {
-  const [searchParams] = useSearchParams();
-  const listParam = searchParams.get("list");
-  return listParam === "LL" ? <Like /> : <PlayListItem /> ;
-};
+import ShortsPlayer from "./pages/ShortsPlayer/ShortsPlayer";
+import VideoPlayer from "./pages/VideoPlayer/VideoPlayer";
+import PlayListPage from "./pages/PlayList/PlayList";
 
 function App() {
   return (
@@ -32,12 +25,24 @@ function App() {
             <Route path={PageEndPoints.RESULTS} element={<ResultsPage />} />
             <Route path={PageEndPoints.CHANNEL} element={<ChannelPage />} />
 
-            <Route path={PageEndPoints.FEEDPLAYLISTS} element={<FeedplayListsPage />}/>
-            <Route path={PageEndPoints.PLAYLIST} element={<PlaylistRoute />} />
-            <Route path={PageEndPoints.FEEDCHANNELS} element={<FeedChannelsPage />}/>
-            <Route path={PageEndPoints.FEEDSUBSCRIPTIONS} element={<FeedSubscriptionsPage />}/>
+            <Route
+              path={PageEndPoints.FEEDPLAYLISTS}
+              element={<FeedplayListsPage />}
+            />
+            <Route path={PageEndPoints.PLAYLIST} element={<PlayListPage />} />
+            <Route
+              path={PageEndPoints.FEEDCHANNELS}
+              element={<FeedChannelsPage />}
+            />
+            <Route
+              path={PageEndPoints.FEEDSUBSCRIPTIONS}
+              element={<FeedSubscriptionsPage />}
+            />
 
-            <Route path={PageEndPoints.SHORTSDETAIL} element={<ShortsPlayer />}/>
+            <Route
+              path={PageEndPoints.SHORTSDETAIL}
+              element={<ShortsPlayer />}
+            />
             <Route path={PageEndPoints.WATCH} element={<VideoPlayer />} />
           </Routes>
         </Router>
