--- src/pages/Feed/PlayLists/PlayListItem/PlayListItem.module.css
@@ -1,168 +1,0 @@
-/*******************************************************/
-/*                                                     */
-/*                        미디어 쿼리                     */
-/*                                                     */
-/*******************************************************/
-@media (min-width: 1075px) {
-    .laterBody{
-        flex-direction: row;
-    }
-    .laterBanner{
-        width: 400px;
-        margin: 20px;
-        height: 90vh;
-        flex-direction: column;
-        border-radius: 10px;
-    }
-    .laterBannerData{
-        flex-direction: column;
-    }
-    .laterBannerMainImg{
-        margin: 0 auto;
-    }
-}
-  
-@media (max-width: 1075px) {
-    .laterBody{
-        flex-direction: column;
-    }
-    .laterBanner{
-        width: 100%;
-        flex-direction: column;
-    }
-    .laterBannerData{
-        flex-direction: row;
-    }
-}
-
-@media (max-width: 780px) {
-    .laterBanner{
-        width: 100%;
-        flex-direction: column;
-    }
-    .laterBannerData{
-        flex-direction: column;
-    }
-    .laterBannerMainImg{
-        margin: 0 auto;
-    }
-}
-
-.laterBody{
-    display: flex;
-}
-/*******************************************************/
-/*                                                     */
-/*                        배너 CSS                      */
-/*                                                     */
-/*******************************************************/
-.laterBanner{
-    padding: 30px;
-    display: flex;
-    align-items: stretch;
-    background-color: rgb(85, 79, 79);
-}
-
-.laterBannerData{
-    display: flex;
-}
-
-.laterBannerMainImg{
-    width: fit-content;
-    padding: 10px;
-    height: 189px;
-    border-radius: 20px;
-}
-
-.laterBannerTitle{
-    font-size: 28px;
-    color: #fff;
-    font-weight: bolder;
-}
-
-.laterBannerUserName{
-    font-size: 14px;
-    color: #fff;
-}
-
-.laterPlaylistInfo span{
-    font-size: 12px;
-    color:#ffffffb3;
-    padding-right: 3px;
-}
-
-.laterBannerBtns{
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    padding: 10px;
-}
-
-.laterBannerBtns button{
-    width: 50%;
-    margin: 0px 10px;
-    min-width: fit-content;
-    height: 36px;
-    border-radius: 18px;
-    border: none;
-    font-size: 14px;
-}
-
-.laterShuffleBtn{
-    background-color: rgba(255,255,255, .3);
-    color:#fff;
-}
-
-/*******************************************************/
-/*                                                     */
-/*                     영상 리스트 CSS                    */
-/*                                                     */
-/*******************************************************/
-.laterPlayList{
-    padding: 10px;
-
-    counter-reset : numbering;
-    list-style-type:none;
-}
-.laterPlayListItem{
-    padding: 8px 0px;
-    display: flex;
-    align-items: center;
-}
-
-.laterPlayList .laterPlayListItem:before{
-    counter-increment : numbering; 
-    content : counter(numbering);
-    margin-right:10px;
-}
-
-.laterPlayListItemImg{
-    width: 200px;
-    height: 113px;
-    border-radius: 10px;
-}
-
-.laterPlayListItemInfo{
-    padding: 0px 8px;
-}
-
-.laterPlayListItemTitle{
-    margin: 0px 0px 8px 0px;
-    font-size: 16px;
-    font-weight: bold;
-}
-
-.laterPlayListItemData span{
-    font-size: 12px;
-    color:#606060;
-    padding-right: 3px;
-}
-
-/*******************************************************/
-/*                                                     */
-/*                        애니메이션                      */
-/*                                                     */
-/*******************************************************/
-.laterBannerBtns button:active{
-    opacity: .7;
-}
