--- src/pages/Like/Like.module.css
@@ -1,112 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: row;
-    gap: 16px;
-    padding: 16px;
-    background-color: #f9f9f9;
-  }
-  
-  .sidebar {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    background-color: #fff;
-    padding: 16px;
-    border-radius: 8px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  }
-  
-  .channelInfo {
-    text-align: center;
-  }
-  
-  .channelThumbnail {
-    width: 120px;
-    height: 120px;
-    border-radius: 50%;
-    margin-bottom: 16px;
-  }
-  
-  .channelTitle {
-    font-size: 20px;
-    font-weight: bold;
-    margin-bottom: 8px;
-  }
-  
-  .channelDescription {
-    font-size: 14px;
-    color: #606060;
-    margin-bottom: 16px;
-  }
-  
-  .playAllButton,
-  .shareButton {
-    width: 100%;
-    padding: 8px 16px;
-    margin-bottom: 8px;
-    border: none;
-    border-radius: 4px;
-    background-color: #065fd4;
-    color: #fff;
-    cursor: pointer;
-  }
-  
-  .videoList {
-    flex: 3;
-    display: flex;
-    flex-direction: column;
-    gap: 16px;
-  }
-  
-  .videoItem {
-    display: flex;
-    gap: 16px;
-    background-color: #fff;
-    padding: 16px;
-    border-radius: 8px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  }
-  
-  .videoThumbnail {
-    width: 160px;
-    height: 90px;
-    border-radius: 8px;
-    object-fit: cover;
-  }
-  
-  .videoDetails {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .videoTitle {
-    font-size: 16px;
-    font-weight: bold;
-    margin-bottom: 8px;
-  }
-  
-  .videoStats {
-    font-size: 14px;
-    color: #606060;
-    margin-bottom: 8px;
-  }
-  
-  .videoDescription {
-    font-size: 14px;
-    color: #606060;
-    display: -webkit-box;
-    -webkit-line-clamp: 2;
-    line-clamp: 2;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-  
-  .noVideos {
-    text-align: center;
-    font-size: 16px;
-    color: #606060;
-  }
-
